requirements.txt
Flask
pycryptodome
pyarmor
pyminifier